components/voice_handler.py
import streamlit as st
from gtts import gTTS
import io
import speech_recognition as sr
import os

## --- 1. Text-to-Speech (TTS) Function ---
def play_audio_button(text, language="ar", unique_key=None):
    """
    Generate karta hai audio aur use ek chote 'Play' button ke saath display karta hai.
    """
    # Unique key zaroori hai Streamlit mein jab multiple buttons hon
    if st.button("🔊 Play", key=f"play_{unique_key}", help="Click to listen to this message"):
        try:
            # 1. gTTS object create karna
            tts = gTTS(text=text, lang=language, slow=False)
            
            # 2. Audio data ko memory mein save karna
            fp = io.BytesIO()
            tts.write_to_fp(fp)
            fp.seek(0)
            
            # 3. Streamlit audio player mein display karna (Hidden audio player)
            # Autoplay=True jab button click hoga.
            st.audio(fp, format='audio/mp3', autoplay=True)
            
            # Message ke neeche ek chota sa indicator
            # st.caption("Playing...") 
            
        except Exception as e:
            st.error(f"Error playing audio: {e}")

## --- 2. Speech-to-Text (STT) Placeholder ---
# (Record and Recognize function jaisa tha, waisa hi rahega)
def record_and_recognize():
    r = sr.Recognizer()
    # ... (baaki code same)
    return r