requirements.txt
streamlit
openai
python-dotenv
langdetect
gTTS
Pillow
pytesseract
opencv-python